position_jitterdodgex {ggtreeExtra}	R Documentation

Simultaneously dodge and jitter, and the whole layer can be shifted vertically or horizontally

Description

This is primarily used for aligning points generated through 'geom_point()' with dodged boxplots (e.g., a 'geom_boxplot()' with a fill aesthetic supplied). And the points can be shifted vertically or horizontally with 'hexpand' or 'vexpand' arguments.

Usage

position_jitterdodgex(
  jitter.width = NULL,
  jitter.height = 0,
  dodge.width = 0.75,
  hexpand = NA,
  vexpand = NA,
  seed = NA
)

Arguments

jitter.width
degree of jitter in x direction. Defaults to 40% of the resolution of the data.

jitter.height
degree of jitter in y direction. Defaults to 0.

dodge.width
the amount to dodge in the x direction. Defaults to 0.75, the default position_dodge() width.

hexpand
The distance to be shifted vertically or horizontally, default is NA.

vexpand
The distance to be shifted vertically or horizontally, default is NA.

seed
A random seed to make the jitter reproducible. Useful if you need to apply the same jitter twice, e.g., for a point and a corresponding label. The random seed is reset after jittering. If NA (the default value), the seed is initialised with a random value; this makes sure that two subsequent calls start with a different seed. Use NULL to use the current random seed and also avoid resetting

[Package ggtreeExtra version 1.4.2 Index]