position_points_sinax {ggtreeExtra}	R Documentation

adjust ridgeline plot position which can be shifted vertically or horizontally.

Description

This is a position adjustment specifically for 'geom_density_ridges()', but it add 'hexpand' that can control shift horizontally.

Usage

position_points_sinax(
  rel_min = 0.02,
  rel_max = 0.98,
  seed = NULL,
  hexpand = NA
)

Arguments

rel_min
numeric, the relative minimum value at which a point can be placed.

rel_max
numeric, the relative maximum value at which a point can be placed.

seed
numeric, Random seed, if set to NULL, the current random number generator is used. If set to NA, a new random random seed is generated. If set to a number, this number is used as seed for jittering only, default is NULL.

hexpand
numeric, distance to be shifted horizontally for geoms that have a position, default is NA.

Value

position method.

Author(s)

Shuangbin Xu

See Also

Other position adjustments: position_dodgex(), position_identityx()

[Package ggtreeExtra version 1.4.2 Index]